Overhauled code to better handle sparse matrices
In the old version of the code, the coordinate matrices were built
as full dense matrices, which ed to unnecessarily large memory
useage when plotting large sparse matrices.  This version reworked
the code so only the relevant coordinate values were stored, reducing
the memory required.

While working on this, I also cleaned up a few bits of sloppy logic
related to 3D arrays and improved the input checking.  I removed the
support for the sz input as a vector... supporting that input was
unnecessarily complicated seeing that a sparse input matrix can do
the same thing (i.e. minimize the memory requirement of the additional
input).